build: rename project and update dependecies to use H2
