--- backend/.run/Run Application.run.xml
@@ -1,17 +1,0 @@
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="Run Application" type="SpringBootApplicationConfigurationType" factoryName="Spring Boot">
-    <module name="productive-areas-api.main" />
-    <option name="SPRING_BOOT_MAIN_CLASS" value="br.com.aegro.productiveareas.ProductiveAreasApplication" />
-    <option name="ALTERNATIVE_JRE_PATH" />
-    <option name="SHORTEN_COMMAND_LINE" value="NONE" />
-    <envs>
-      <env name="SPRING_DATASOURCE_URL" value="jdbc:postgresql://localhost:5432/productive-areas-db" />
-      <env name="SPRING_DATASOURCE_USERNAME" value="postgres" />
-      <env name="SPRING_DATASOURCE_PASSWORD" value="postgres" />
-      <env name="SPRING_JPA_HIBERNATE_DDL_AUTO" value="update" />
-    </envs>
-    <method v="2">
-      <option name="Make" enabled="true" />
-    </method>
-  </configuration>
-</component>
