--- backend/docker-compose.yml
@@ -1,23 +1,0 @@
-version: '3.9'
-
-services:
-  app:
-    image: 'productive-areas-api:latest'
-    build:
-      context: .
-    container_name: app
-    depends_on:
-      - db
-    environment:
-      - SPRING_DATASOURCE_URL=jdbc:postgresql://db:5432/productive-areas-db
-      - SPRING_DATASOURCE_USERNAME=postgres
-      - SPRING_DATASOURCE_PASSWORD=postgres
-      - SPRING_JPA_HIBERNATE_DDL_AUTO=update
-
-  db:
-    image: 'postgres:alpine'
-    container_name: db
-    environment:
-      - POSTGRES_DB=productive-areas-db
-      - POSTGRES_USER=postgres
-      - POSTGRES_PASSWORD=postgres
