+++ backend/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'teste-backend'
+rootProject.name = 'fms'
