+++ backend/src/main/java/br/com/aegro/fms/ServletInitializer.java
@@ -1,0 +1,11 @@
+package br.com.aegro.fms;
+
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+public class ServletInitializer extends SpringBootServletInitializer {
+	@Override
+	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+		return application.sources(FarmManagementSystemApplication.class);
+	}
+}
