+++ backend/src/main/resources/application.properties
@@ -1,1 +1,1 @@
-spring.application.name=productive-areas-api
+spring.application.name=farm-management-system
