+++ backend/src/test/java/br/com/aegro/fms/FarmManagementSystemApplicationTests.java
@@ -1,10 +1,10 @@
-package br.com.aegro.productiveareas;
+package br.com.aegro.fms;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class ProductiveAreasApplicationTests {
+class FarmManagementSystemApplicationTests {
 
 	@Test
 	void contextLoads() {
